+++ src/exceptions/NonReversibleCommandException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace yentu\exceptions;
+
+
+class NonReversibleCommandException extends \Exception
+{
+
+}